Create list manager
Realizing direct DOM manipulations are going to grow tedious as the app
grows, I added a render method that calls uhtml. The elements will
remain interoperable as they're still web components.

Had trouble passing down the property with uhtml because it set the prop
before the element was constructed. That caused prop setter function to
no longer be called due to prototypical inheritance. That is why there
is a `delete` call in the AccountList constructor.